Entire project update
I have not been uploading this to github. Currently, it takes in the
spectator data from Riot's spectator client and parses out the summoner
data from keyframes. The summoner data is only the first ~20% of the
data. Next step would ideally be turrets and from there the rest of the
data.